### Speedy Concurrent Cluster Login with Osprey - A Game Changer!

🎉 **Introducing the Revolutionary Concurrent Cluster Login Feature with Osprey!** 🎉

Do you find yourself waiting forever as Osprey logs you into each cluster one by one? Are you tired of wasting precious time watching cluster names appear on the screen, one after the other? Well, we have fantastic news for you! Our latest feature brings the power and speed of concurrent logins to Osprey, transforming your experience from slow and tedious to lightning-fast and efficient!

🔧 **What's New?**
- **Concurrent Logins**: No more watching cluster names appear one by one. Now, Osprey logs you into multiple servers at the same time, slashing your wait time significantly.
- **Supercharged Speed**: Experience the thrill of near-instantaneous logins. Your workflow will be smoother and more productive than ever before!
- **Effortless Integration**: This feature seamlessly integrates into your current setup. No complicated configurations or additional tools needed. Just plug and play!

💡 **Why You Need This Feature**:
- **Save Time**: Time is money! Reduce your login time from minutes to seconds.
- **Boost Productivity**: Spend less time waiting and more time doing what you do best.
- **Simplified Workflow**: Make your process more efficient and less frustrating.

📈 **Results That Speak for Themselves**:
Imagine logging in to all your clusters in a fraction of the time it used to take. Our tests show an incredible improvement in speed and performance, making your cluster management tasks a breeze.

```shell
# before
osprey user login --group non-prod  0.95s user 0.17s system 14% cpu 7.628 total
# after
./cosprey user login --group non-prod  0.90s user 0.14s system 30% cpu 2.377 total
```

🔜 **Coming Soon to a Repository Near You!**
Don't miss out on this incredible upgrade. Update your repository now and experience the future of cluster logins with Osprey!

🎊 **Upgrade Today and Watch Your Productivity Soar!** 🎊

**Order your update now by merging this PR and say goodbye to slow, sequential logins forever!**

---

**PR Includes**:
- Implementation of concurrent cluster login functionality in Osprey.
- Refactoring of OIDC library to hide logic which was exposed unnecessarily in Azure implementation.
- No updates to documentation as this feature is user-transparent.

---

**Experience the magic of concurrent logins with Osprey. Merge this PR and turbocharge your workflow today!**

---

*Note: Results may vary based on the number of clusters and network conditions. For best results, ensure your environment is properly configured for concurrent operations.*