@aecay 2024-07-08
awesome work.  there may be further speed gains to be had as well.

currently we have an api which is `RetrieveClusterDetailsAndAuthTokens`.  that does two things:
- wait for the oidc flow to happen in the browser
- download some data (notably the cluster ca fingerprint) from an endpoint, per cluster

we should be doing the second step in the background while we are waiting for the browser.  if we do that, then the time taken for `osprey user login` should be dominated by the time the oidc flow takes.

-----

on the code in this PR, though, you're calling the azure implementation of [`RetrieveClusterDetailsAndAuthTokens`](https://github.com/sky-uk/osprey/blob/5182564842bda2a58df5b01d9619dadbf7fefd7a/client/azure.go#L138) in parallel in goroutines.  i'm not convinced that that's goroutine-safe.  won't that try to call [`r.oidc.AuthWithOIDCCallback`](https://github.com/sky-uk/osprey/blob/5182564842bda2a58df5b01d9619dadbf7fefd7a/client/azure.go#L147) multiple times in parallel?

(we're also not doing context propagation, so an error in one cluster fecth won't abort the other clusters.  that may or may not be what we want; certainly in today's sequential world we try all clusters rather than stopping on the first error.  what i'm slightly more worried about is signal handling, and whether ctrl-c will successfully terminate the program [all goroutines].  but i'm rusty on how signal handling works in go....)