@zemanlx 2024-07-08
Thank you @aecay!

I spent some time with the method `RetrieveClusterDetailsAndAuthTokens` and it seems to be complicated. I would rather have two methods for being responsible for one job only in Azure, but it is needed for the Osprey server where it is all done in one step. If we have separate methods we can do OIDC flow concurrently while we fetch all information from clusters. So unless we want to change interface/API there is no easy way to do it I can think of. 

Right now we are waiting for the first cluster to finish using [mutex](https://github.com/sky-uk/osprey/pull/106/files#diff-81bc9624b349e316cf34a69d957548920820fd95a5bc9a22e9c694c08fe8e3dcR68-R69) and then reuse that token for all other clusters (similar to what was happening when we were calling it sequentially). But because of that mutex, it is safe as no two goroutines can get token/start OIDC flow concurrently. However, they can concurrently ask for server info.

I am doing context propagation as far as it is supported. However, I see your point that the way it is done does not make much sense. If we want to make it meaningful we would need to change the interface of `retriever` to support passing context. If you are ok with that, I can do it, but it will be much more lines to review 🙈