@aecay 2024-07-08
> it is needed for the Osprey server where it is all done in one step

i'm pretty sure nothing uses osprey server.  as in, nothing at all.  so we could get rid of it (in a major version upgrade of osprey) if that would make our life easier.

> If you are ok with [a different approach to context propagation], I can do it, but it will be much more lines to review

nah, don't bother.  i wanted to talk about the complexities and limitations of the code (e.g. if you were going to be doing further work on the codebase, to make sure you took them into account), but i don't want to cause extra work or bloat the scope of this PR.

> Right now we are waiting for the first cluster to finish using [mutex](https://github.com/sky-uk/osprey/pull/106/files#diff-81bc9624b349e316cf34a69d957548920820fd95a5bc9a22e9c694c08fe8e3dcR68-R69)

thanks -- i missed the subtlety of how that mutex fitted into the flow.  now i see how the code is goroutine-safe.